Fix UI related to permissions
