+++ src/ui/options-permissions.js
@@ -1,19 +1,15 @@
 import Settings from '../lib/settings.js';
+import * as lib from './lib.js';
 
 /** @type {Map<String,"yes"|"no"|"unavailable">} */
-const permissionStatuses = new Map();
+let permissionStatuses = new Map();
 
 async function loadPermissions() {
-  /** @type {Map<String,"yes"|"no"|"unavailable">} */
-  await Promise.all(['bookmarks', 'tabs', 'tabGroups'].map(async (perm) => {
-    try {
-      const status = await browser.permissions.contains({ permissions: [perm] });
-      permissionStatuses.set(perm, status ? 'yes' : 'no');
-    } catch (e) {
-      permissionStatuses.set(perm, 'unavailable');
-    }
-  }));
+  permissionStatuses = await lib.loadPermissions();
+}
 
+function refreshUi() {
+  lib.hideUiIfPermissionsNotGranted(permissionStatuses);
   document.querySelectorAll('[data-request-permission]').forEach((el) => {
     const status = permissionStatuses.get(el.dataset.requestPermission);
     if (status === 'unavailable') {
@@ -51,32 +47,21 @@ async function loadPermissions() {
       el.classList.add('is-hidden');
     }
   });
-
-  document.querySelectorAll('[data-hide-if-permission-contains]').forEach((el) => {
-    const status = permissionStatuses.get(el.dataset.hideIfPermissionContains);
-    if (status === 'unavailable') {
-      // eslint-disable-next-line no-param-reassign
-      el.innerText = 'Unsupported';
-    } else if (status === 'yes') {
-      // eslint-disable-next-line no-param-reassign
-      el.classList.add('is-hidden');
-    } else if (status === 'no') {
-      el.classList.remove('is-hidden');
-    }
-  });
-
-  document.querySelectorAll('[data-dependson-permissions]').forEach((el) => {
-    const dependsOn = el.dataset.dependsonPermissions.split(',');
-    const statuses = dependsOn.map((perm) => permissionStatuses.get(perm));
-    // eslint-disable-next-line no-param-reassign
-    el.disabled = !statuses.every((dep) => dep === 'yes');
-  });
 }
 
-document.addEventListener('DOMContentLoaded', loadPermissions);
+document.addEventListener('DOMContentLoaded', async () => {
+  await loadPermissions();
+  refreshUi();
+});
 
-browser.permissions.onAdded.addListener(loadPermissions);
-browser.permissions.onRemoved.addListener(loadPermissions);
+browser.permissions.onAdded.addListener(async () => {
+  await loadPermissions();
+  refreshUi();
+});
+browser.permissions.onRemoved.addListener(async () => {
+  await loadPermissions();
+  refreshUi();
+});
 
 document.querySelectorAll('[data-request-permission]').forEach((node) => {
   node.addEventListener('click', async (e) => {
